minor 1.5 fix.

+++ astrogrid/desktop/impl/src/java/org/astrogrid/desktop/modules/ui/scope/ScopeServicesList.java
@@ -183,7 +183,7 @@ public class ScopeServicesList extends RegistryGooglePanel
                     } else if (arg1 instanceof String) {
                         return 1;
                     } else {
-                        return ((Integer)arg0).compareTo(arg1);
+                        return ((Integer)arg0).compareTo((Integer)arg1);
                     }
                 }
             }){
